genaral cohesive energy evaluation
When cohesive energy convergence test an F(fluorine) use structure SiF4,
automatically setting by the convergence workflow. Similar of
lanthanides, use Re-N.

Things become tricky, when the isolate atom energy of lanthanides is
going to be calculated. NEEDED TO increase the `nbnds` and use much more
RAM(by in crease the number of machines).

The inputs parameters of convergence workflow are `ecutwfc_list` and
`ecutrho_list` with dual concept deprecated.