Replace F77 efermi with pure python efermi
Build fortran code bring some unconvinients to package maintainance such
as failed in building docs in readthedocs since the docker image of
readthedocs contain no fortran compiler, and the problem of wired
setup.py setting to build and install the package.

I copy and paste the pure python efermi code from Austin to the package.
The performence not a big concern in this sssp-workflow, because it only
used in calculating the fermi level for metals and fermi level with
specific energy level shifts.