Replay all convergence with legacy version
The previous implementation of convergence workflow use aiida-optimize
as the convergence engine, which is advanced and will be the future
practical implementation. However, it faces two problem, 1) the
aiida-optimize is under development and has no enough users. The
fail-continue issues still not resolved. 2) The concept of using
convergence engine to finish the convergence test right after the
covergence criterien reach is different from the legacy implementation,
therefore the convergence parameters such as tolerance and
convergence_window are hard to set.

The legacy convergence workflows are construct to perfectly reproduce
the results of SSSP results publish now at MaterialCloud. Besides it is
also useful in compare with the new logic convergence workflow(e.g
aforementioned aiida-optimize engine one) as the benchmark.

The following legacy convergence workflows are completed in this commit:
- cohesive energy
- pressure
- phonon frequencies
- bands distance

Meanwhile, there is a `BandsDistanceWorkChain` for calculating the
distances between each pairs of given pseudopotentials list. It will
give all pp pairs bands distance as outputs which can further used for
bands distance chessboard generation.