Protocol: separate type of protocols and add new protocol
There are suppose to be two kinds of protocols, one called calculation
protocol, which are the input parameters for different verification
workflows to run, it will influence the result of the calculation.
The other type of protocol is the selection protocol. It act on the
inspection and selection process to decide which pseudopotential is good
to choose in to the SSSP library and which pseudopotential should be
regard as converge.

Now in the repository, there is only the calculation protocol set in the
YAML file CALC_PROTOCOL.yml.