Major update: Convergence on wfc and rho
In this PR, the dual concept is totally deprecated from the users view.
The dual only used in the inner workflow to control the rho cutoff
setting.
The criteria protocol is added to measure when should the  convergence
stop. All convergence workflows are refactoring where the complexities
are included in the _base convergenece workflow and the sub workflows
are only for properties related parameters setting.