Fix: class variable all move to ctx (#91)
fixes #89

Some class variables such as self._REFERENCE_ECUTWFC in _base convergence work chain is not stored in the ctx of work chain. It is fine if the daemon is not restarted but leads to a problem that the class lost this variable when the daemon restart. Since the variable is not set with the class but in the step.

To resolve this, all the variables that pass to other work chain steps need to be stored as ctx.

- also remave all _MAX_WALLTIME class attributes.