Ph calculation sure the remote path is there
The caching mechanism is tricky for the phonon calculation.
See discussion https://github.com/aiidateam/aiida-core/issues/5178
It is workaround by before running ph calculation, always check the
remote path of the previous pw calculation. If it is empty means the
folder is cleaned therefore potentially will fail the PH calculation.
And the pw calculation will actually run without from using caching
node.

It is not a perfect solution since even if the pw calculation workdir is
being cleaned, if the subseqent ph calculation is succeed and used as
caching node, in principle the pw calculation no need to run from
scratch without caching.