Able to set wavefunction cutoff and skip to rho test only (#123)
Setting preset_check for convergence workflow will skip wavefunction cutoff test and only process the rho test.
This is useful for branching the tests for different criteria. When running either convergence test on efficiency or precision criteria, the wfc test is the same. This step will write recommended cutoff of all criteria to the output.
When rerun the convergence test the preset_check can be used to run on that cutoff for rho cutoff test only.