Set dual depend on elements and confs in delta and bands measure (#127)
For different elements and configurations, in delta accuracy measure and bands measure, the dual should be different. 
`dual=8` was used which not proper. Now for oxides, non-NC dual equals to 8. For NC, dual equal to 4 for unarias and typical configurations. For high dual element, dual is set to 18 which is very high to make sure charge density cutoff is proper.